feat: 4차 디자인 수정 추가
